+++ template_test.go
@@ -68,6 +68,7 @@ func TestEvalUseStmt(t *testing.T) {
 		{"8.with-component", nil},
 		{"9.with-inserts-and-html", nil},
 		{"11.with-component-no-args", nil},
+		{"10.with-component-and-slots", nil},
 	}
 
 	tpl, err := NewTemplate(&Config{